requirements.txt
python-dotenv==1.0.0
pypdf==3.12.0
langchain==0.0.252
openai==0.27.8
faiss-cpu==1.7.4
uvicorn==0.22.0
fastapi==0.98.0
jinja2==3.1.2
unstructured==0.8.1
tiktoken==0.4.0
sentence-transformers==2.2.2
pytest==7.4.0
httpx==0.24.1
python-multipart==0.0.6
gpt4all==1.0.8
InstructorEmbedding==1.0.1
